+++ PacDefence/src/sprites/Pacman.java
@@ -37,7 +37,7 @@ public class Pacman extends AbstractSprite {
    private static final List<BufferedImage> images = Collections.unmodifiableList(makeImages());
    
    
-   public Pacman(int currentLevel, int hp, List<Point> path) {
+   public Pacman(int currentLevel, long hp, List<Point> path) {
       super(images, currentLevel, hp, path);
    }
